Refactoring: isolated different committing strategies in order to simplify the code
