+++ package.yaml
@@ -1,5 +1,5 @@
 name: nakadi-client
-version: '0.5.1.0'
+version: '0.6.0.0'
 synopsis: Client library for the Nakadi Event Broker
 description: This package implements a client library for interacting
              with the Nakadi event broker system developed by Zalando.
